some parts of validations in services didn't working, they are now fixed(check capacity in updating adding connectors)

+++ src/Domain/Services/ConnectorService.cs
@@ -1,5 +1,7 @@
+using AutoMapper;
 using Domain.Contracts.Repository;
 using Domain.Contracts.Services;
+using Domain.DTOs;
 using Domain.Entities;
 
 namespace Domain.Services;
@@ -8,46 +10,56 @@ public class ConnectorService : IConnectorService
 {
     private readonly IConnectorRepository _repository;
     private readonly IGroupRepository _groupRepository;
+    private readonly IMapper _mapper;
 
-    public ConnectorService(IConnectorRepository repository, IGroupRepository groupRepository)
+    public ConnectorService(IConnectorRepository repository, IGroupRepository groupRepository, IMapper mapper)
     {
         _repository = repository;
         _groupRepository = groupRepository;
+        _mapper = mapper;
     }
 
-    public async Task<IEnumerable<Connector>> GetAllAsync()
+    public async Task<IEnumerable<ConnectorDTO>> GetAllAsync()
     {
-        return await _repository.GetAllAsync();
+        var result = await _repository.GetAllAsync();
+        return _mapper.Map<List<ConnectorDTO>>(result);
     }
 
-    public async Task<Connector> CreateAsync(Connector connector)
+    public async Task<ConnectorDTO> CreateAsync(ConnectorDTO connector)
     {
         if (connector.ChargeStationId == 0)
             throw new ArgumentException("A Connector cannot exist in the domain without a Charge Station.");
 
         var group = _groupRepository.GetByStationIdAsync(connector.ChargeStationId);
+        var groupChargeStationCount = await _groupRepository.GetChargeStationsCountAsync(group.Id);
+        if (groupChargeStationCount >= 5)
+            throw new Exception("Maximum charge stations for a group is 5, so you can not add more charge stations.");
 
-        if (group.Capacity < group.ChargeStations.Sum(c => c.Connectors.Sum(x => x.MaxCurrent)))
-            throw new ArgumentException(
-                "The capacity of the Group is not enough to add new connector");
+        var groupChargeStationCurrent = await _groupRepository.GetChargeStationsCurrentAsync(group.Id);
+        if (group.Capacity < groupChargeStationCurrent + connector.MaxCurrent)
+            throw new ArgumentException("The capacity of the Group is not enough to add new connector");
 
-        return await _repository.CreateAsync(connector);
+        var result = await _repository.CreateAsync(_mapper.Map<Connector>(connector));
+        return _mapper.Map<ConnectorDTO>(result);
     }
 
-    public async Task<Connector> UpdateAsync(Connector connector)
+    public async Task<ConnectorDTO> UpdateAsync(ConnectorDTO connector)
     {
-        var data = await _repository.GetConnectorIncludingGroupGroup(connector.Id);
-        if (data == null)
+        var connectorFromDb = await _repository.GetConnectorIncludingGroup(connector.Id);
+        if (connectorFromDb == null)
             throw new ArgumentException($"Connector with Id{connector.Id} not found.");
         if (connector.ChargeStationId == 0)
             throw new ArgumentException("A Connector cannot exist in the domain without a Charge Station.");
 
-        var group = data.ChargeStation.Group;
-        if (group.Capacity < group.ChargeStations.Sum(c => c.Connectors.Sum(x => x.MaxCurrent)))
+        var group = connectorFromDb.ChargeStation.Group;
+        var groupChargeStationCurrent = await _groupRepository.GetChargeStationsCurrentAsync(group.Id);
+        if (group.Capacity < groupChargeStationCurrent + connector.MaxCurrent)
             throw new ArgumentException(
-                "The capacity in Amps of a Group should always be great or equal to the sum of the Max current in Amps of the Connector of all Charge Stations in the Group.");
+                "The capacity of the Group is not enough to update the connectors max current.");
 
-        return await _repository.UpdateAsync(connector);
+        connectorFromDb.MaxCurrent = connector.MaxCurrent;
+        var result = await _repository.UpdateAsync(connectorFromDb);
+        return _mapper.Map<ConnectorDTO>(result);
     }
 
     public async Task RemoveAsync(long id)
@@ -55,6 +67,11 @@ public class ConnectorService : IConnectorService
         var connector = await _repository.GetAsync(id);
         if (connector == null)
             throw new ArgumentException($"Connector with Id{id} not found.");
+
+        var group = _groupRepository.GetByStationIdAsync(connector.ChargeStationId);
+        var groupChargeStationCount = await _groupRepository.GetChargeStationsCountAsync(group.Id);
+        if (groupChargeStationCount <= 1)
+            throw new ArgumentException("Charge station should have at least one connector.");
 
         await _repository.RemoveAsync(connector);
     }
